+++ app/User/repository.js
@@ -3,24 +3,24 @@
  */
 const User = require('./model')
 
-const get = () =>
-  User.fetchAll()
+const find = query =>
+  User.where({...query}).fetchAll()
 
-const byId = id =>
-  new User({id}).fetch()
+const findOne = query =>
+  User.where(query).fetch()
 
 const create = (name, email) =>
   new User({name, email}).save()
 
 const update = (id, update) =>
-  byId(id).then(user => user.set(update).save())
+  findOne({id}).then(user => user.set(update).save())
 
 const remove = id =>
-  byId(id).then(user => user.destroy({hardDelete: true}))
+  findOne({id}).then(user => user.destroy({hardDelete: true}))
 
 module.exports = {
-  get,
-  byId,
+  find,
+  findOne,
   create,
   remove,
   update